Map multi match to new action
