+++ Android/app/src/androidTest/java/uk/ac/lshtm/keppel/android/MatchActionTest.kt
@@ -157,29 +157,4 @@ class MatchActionTest {
             equalTo(96.0)
         )
     }
-
-    @Test
-    fun whenPassedAListOfTemplates_capturesAndReturnsMaxMatchScore() {
-        fakeMatcher.addScore("blah1", "scanned", 94.0)
-        fakeMatcher.addScore("blah2", "scanned", 96.0)
-        fakeMatcher.addScore("blah3", "scanned", 95.0)
-
-        val intent = Intent(External.ACTION_MATCH).also {
-            it.putExtra(OdkExternal.PARAM_INPUT_VALUE, "foo")
-            it.putExtra(External.paramIsoTemplate(1), "blah1".toHexString())
-            it.putExtra(External.paramIsoTemplate(2), "blah2".toHexString())
-            it.putExtra(External.paramIsoTemplate(3) + "_3", "blah3".toHexString())
-        }
-
-        val result = rule.launchAction(intent, ConnectingPage()) {
-            it.connect(fakeScanner, MatchPage()).clickMatch()
-            fakeScanner.returnTemplate("scanned", 1)
-        }
-
-        val extras = result.resultData.extras!!
-        assertThat(
-            extras.getDouble(OdkExternal.PARAM_RETURN_VALUE),
-            equalTo(96.0)
-        )
-    }
 }
