+++ Android/app/src/main/AndroidManifest.xml
@@ -40,6 +40,11 @@
                 <action android:name="uk.ac.lshtm.keppel.android.MATCH" />
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
+
+            <intent-filter>
+                <action android:name="uk.ac.lshtm.keppel.android.MULTI_MATCH" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
         </activity>
 
         <activity android:name="uk.ac.lshtm.keppel.android.settings.UsbDeviceAttachedActivity"
