+++ Android/app/src/main/java/uk/ac/lshtm/keppel/android/External.kt
@@ -5,6 +5,7 @@ object External {
 
     const val ACTION_SCAN = "$APP_ID.SCAN"
     const val ACTION_MATCH = "$APP_ID.MATCH"
+    const val ACTION_MULTI_MATCH = "$APP_ID.MULTI_MATCH"
 
     const val PARAM_RETURN_ISO_TEMPLATE = "$APP_ID.return_iso_template"
     const val PARAM_RETURN_NFIQ = "$APP_ID.return_nfiq"
